py/objstr.c: Add new mp_obj_new_str_0() method.
There were lots of places where this pattern was duplicated, to convert a standard C string to a MicroPython string:

x = mp_obj_new_str(s, strlen(s));

This commit provides a simpler method that removes this code duplication:

x = mp_obj_new_str_0(s);

The _0 suffix is because it accepts a NUL-terminated string as a parameter.

This gives clearer, and probably smaller, code.

Signed-off-by: Jon Foster <jon@jon-foster.co.uk>