docs/library/time.rst: Document the minimum value of TICKS_PERIOD.
The old documentation about ticks was not enough to do anything useful
with them.  They might have overflowed after 2 milliseconds, according
to the documentation, which is too short for them to be useful.
Document what the actual minimum is.  Unsurprisingly, it's large enough
to allow useful work to be done with these timers.

Signed-off-by: Jon Foster <jon@jon-foster.co.uk>