py/obj.h: Add new int constructors for 32-bit integers
Add mp_obj_new_int_from_int32 and mp_obj_new_int_from_uint32.

These have the potential to be faster and simpler than mp_obj_new_int()
or mp_obj_new_int_from_uint() on 64-bit platforms, because the compiler
can deduce that the argument will definitely fit a small int, and it
can optimize accordingly.

On 32-bit platforms, these are the same as mp_obj_new_int() and
mp_obj_new_int_from_uint().  Right now I'm counting on the compiler or
linker to notice that and optimize accordingly.  These could be made
into aliases in the header file in that case, but I couldn't figure out
the appropriate ifdefs, so that is left for a possible future
improvement.

Signed-off-by: Jon Foster <jon@jon-foster.co.uk>