shared/timeutils/timeutils.h: Add TIMEUTILS_SECONDS_FROM_1970_TO_EPOCH.
It's very difficult to write C code that deals with time without knowing
what the epoch is.  Some of the existing MicroPython code has just been
copy-pasted twice for the two epochs that are currently supported.

This adds a new define, TIMEUTILS_SECONDS_FROM_1970_TO_EPOCH, which is
defined in timeutils.h.  This is 0 if the epoch is 1970, or the relevant
offset otherwise.  Time code can just add or subtract this value, and it
will work for any epoch.  Two of the functions in timeutils.h have been
de-duplicated in this way.

Signed-off-by: Jon Foster <jon@jon-foster.co.uk>